move age check into file
